frontend/src/app/auth/auth.guard.ts
import { AuthService } from '../services/AuthService';
import { Injectable } from '@angular/core';
import { CanActivate, Router } from '@angular/router';
import { Observable } from 'rxjs';
import { map, tap, take } from 'rxjs/operators';

@Injectable({
  providedIn: 'root'
})
export class AuthGuard implements CanActivate {

  constructor(private authService: AuthService, private router: Router) { }

  canActivate(): Observable<boolean> | Promise<boolean> | boolean {
    return this.authService.isAuthenticated().pipe(
      take(1),
      map(isAuthenticated => {
        if (isAuthenticated) {
          return true; // Usuário autenticado, permite acesso
        } else {
          this.router.navigate(['/login']); // Redireciona para a página de login
          return false; // Bloqueia o acesso
        }
      })
    );
  }
}
